pyproject.toml
[project]
name = "docx-revisions"
version = "0.1.1"
description = "Track changes (revisions) support for python-docx: read, write and manipulate <w:ins> and <w:del> elements"
readme = "README.md"
license = { text = "MIT" }
license-files = ["LICEN[CS]E*"]
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.urls]
Homepage = "https://github.com/balalofernandez/docx-revisions"
Documentation = "https://balalofernandez.github.io/docx-revisions"
Repository = "https://github.com/balalofernandez/docx-revisions"

[tool.ruff]
exclude = []
extend-exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".github",
    ".mypy_cache",
    "venv*",
    "docs",
]
line-length = 120

[tool.ruff.format]
skip-magic-trailing-comma = true
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "A",
    "B",
    "E1",
    "E4",
    "E7",
    "E9",
    "F",
    "I",
    "PLE",
    "PLW",
    "RUF100",
    "SIM",
    "UP",
    "W",
]
ignore = [
    "PLW0603",
    "UP006",
    "UP007",
    "UP035",
    "A002",
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["docx_revisions", "tests"]
split-on-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_classes = ["Describe*", "Test*"]
python_functions = ["it_*", "test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
